Force removal of piHoleVersion to avoid write-protection warning
When starting up v1.3, my chronometer2 would detect the piHoleVersion file and try to remove it, but the file is write-protected so the script would require confirmation before continuing. This isn't great for headless mode, so I've added the -f flag to the file removal to force removal regardless of write status.