Update date checking fix
Correct the behavior of the version checking function to update today's date before comparing it to the date of last check. 

This fixes the existing behavior of: if the date the script is started is the same as the date of last check (e.g. you performed a quick restart of your pihole but a check happened earlier that day), then the GetVersionInformation function would never reacquire today's date nor actually check version information. This will also indirectly fix the issue of PADD returning out-of-date version messages indefinitely (unless the script is run again) if run on startup before the pihole DNS is fully initiated, since version checking will now actually happen ;)