Use ISO 6429 SGR parameters to set text attributes
For reference, see -
https://en.wikipedia.org/wiki/ANSI_escape_code#SGR_(Select_Graphic_Rendition)_parameters

The escape sequences for some terminals (e.g. rxvt-unicode) contain
semicolons in them. E.g.  `tput setaf 0` returns "[38;5;0m"

Now, we set heatmap colors using these escape sequences, and we store
the heatmap in ./piHoleVersion without quoting it.  So, when we source
./piHoleVersion, bash interprets the semicolons as command-separators,
and shows a bunch of errors like these -

	piHoleVersion: line 4: 5: command not found
	piHoleVersion: line 4: 2m: command not found
	piHoleVersion: line 7: 5: command not found
	piHoleVersion: line 7: 1m: command not found
	piHoleVersion: line 10: 5: command not found
	piHoleVersion: line 10: 2m: command not found
	piHoleVersion: line 13: 5: command not found
	piHoleVersion: line 13: 2m: command not found
	piHoleVersion: line 15: 5: command not found
	piHoleVersion: line 15: 2m: command not found

for reference, these are the contents of line 4, 7, 10, 13 and 15 -

	core_version_heatmap=^[[38;5;2m
	web_version_heatmap=^[[38;5;1m
	ftl_version_heatmap=^[[38;5;2m
	padd_version_heatmap=^[[38;5;2m
	version_heatmap=^[[38;5;2m

So now, as suggested by @dschaper in [1], we use hardcoded ANSI escape
codes that are guaranteed to work with ISO 6429 compatible terminals.

[1]: https://github.com/pi-hole/PADD/pull/171#issuecomment-902009001,